created separate conftest for webtests
